requirements.txt
fastapi
pydantic
numpy
matplotlib
pytesseract
pillow
starlette
opencv-python